Add workspaceApp.addLink
